requirements.txt
# --- Core Django ---
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==4.1

# --- Base de datos ---
psycopg2-binary==2.9.10  # PostgreSQL

# --- Análisis y soporte ---
pandas==2.3.0
openpyxl==3.1.5        # Excel
pillow==10.4.0         # Imágenes en Django
python-decouple==3.8   # Manejo de configuración (opcional, si lo usas)

# --- Dependencias secundarias (instaladas automáticamente, no siempre necesarias fijar) ---
asgiref==3.8.1
sqlparse==0.5.1
tzdata==2025.2
python-dateutil==2.9.0.post0
pytz==2024.2
